feat: 3 parallel decision agents + natural demo-mode violation injection
- 决策阶段改为 3 个并行 Agent（价值投资/趋势交易/综合决策），
  结果筛选优化只对三个候选排序推荐
- compliance_checker 从图节点改为后置函数 review_compliance
- 演示模式：B1 prompt 引导激进输出 + B2 自然激进话术兜底注入
- db/session 加 pool_pre_ping 修复 Neon 空闲连接回收导致的 500

+++ demo_service/schemas.py
@@ -22,7 +22,7 @@ class AdvisoryReport(BaseModel):
     risk_report: str = ""
     decision: str = ""
     compliance_result: str = ""
-    compliance_score: int | None = None
+    compliance_score: float | None = None
     final_output: str = ""
     violations: list[ViolationItem] = []
